This is a tricky problem. I solved it by just looking for the pattern for hours. Mental brute force lol. The pattern can be simply explaned tho by significant digits:
238798321
|       ^-least significant decimal digit
^-most significant digit

Find the least significant digit that is immediately followed by a digit lower than it.

238798321
   |^- 9 - least significant decimal digit where the next most significant is a smaller digit
   ^- 7 - next most significant digit after 9

This identifies a range we're interested int. We'll call them the misplaced digits

238798321
238 - correctly placed, ignore
798321 - "misplaced"

To rearrange the digits in "misplaced" so they will provide the next greater element with the same digits (our solution), isolate an instance of the digit in misplaced that is the next largest digit value above misplaced's most significant digit.

798321
| ^- 8 - the digit in misplaced that is the next largest digit value above misplaced's most significant digit.
^- 7 - misplaced's most significant digit

The digit in misplaced that is the next largest digit value above misplaced's most significant digit will now be the most significant digit among the set of correctly placed digits. The rest are just the remaining digits sorted so the higher digits occupy the least significant decimal places.

8 - isolated, to be most significant digit of "correct placed" digits
12379 - the rest of the correctly placed digits, sorted

Concatencate 238 with 812379 to get the solution 238812379