Make a scanning row (r) that you will compare with each row in the matrix. This scanning row tracks the "streak" for each column (how many previous rows had the bit set in that column). If a row does has a zero for a column the scanning row at that column position gets reset to zero, otherwise the streak count at that column on the scanning row is incremented.
After the scanning row is compared with a row, a copy of the scanning row is made and sorted so large values come first (this is equvialent to reordering the columns). Then the value at each row is multiplied by its index plus one. This operation sizes the submatrix and we greedily store the submatrix with the largest size.